Add files via upload
The raw are provided in the supplementary files 1 and 2. The raw data is showing the time series data namely Relative Humidity and Temperature. The maximum reading for daily Relative Humidity is 98.1% while the minimum is 69.5%. As for daily Temperature, the maximum is 29.5°C while the minimum is 23.7°C. Both Relative Humidity and air Temperature intercorrelated with each other.